feat: refine the search bar UX
