+++ lib/screens/sponsor/sponsor_screen.dart
@@ -1,28 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:suica_no_toy_box_flutter/constants/dimensions.dart';
 import 'package:suica_no_toy_box_flutter/cubits/sponsor/sponsor_cubit.dart';
 import 'package:suica_no_toy_box_flutter/cubits/sponsor/sponsor_state.dart';
 import 'package:suica_no_toy_box_flutter/screens/sponsor/company_detail_card.dart';
 import 'package:suica_no_toy_box_flutter/screens/sponsor/company_list.dart';
+import 'package:suica_no_toy_box_flutter/screens/sponsor/company_search_field.dart';
 
-class SponsorScreen extends StatefulWidget {
+class SponsorScreen extends StatelessWidget {
   const SponsorScreen({super.key});
-
-  @override
-  State<SponsorScreen> createState() => _SponsorScreenState();
-}
-
-class _SponsorScreenState extends State<SponsorScreen> {
-  final _formKey = GlobalKey<FormBuilderState>();
-
-  void _onSearchSubmitted(String? value) {
-    if (value != null && value.isNotEmpty) {
-      context.read<SponsorCubit>().searchCompany(value);
-    }
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -71,35 +57,7 @@ class _SponsorScreenState extends State<SponsorScreen> {
             spacing: 16,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              FormBuilder(
-                key: _formKey,
-                child: FormBuilderTextField(
-                  name: 'company_search',
-                  decoration: InputDecoration(
-                    hintText:
-                        'Search for company sponsorship information in the UK',
-                    prefixIcon: const Icon(Icons.search),
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(8.0),
-                    ),
-                    filled: true,
-                    fillColor: Theme.of(context).colorScheme.surfaceContainer,
-                    suffixIcon: IconButton(
-                      icon: const Icon(Icons.clear),
-                      onPressed: () {
-                        _formKey.currentState?.fields['company_search']
-                            ?.reset();
-                        context.read<SponsorCubit>().searchCompany('');
-                      },
-                    ),
-                  ),
-                  validator: FormBuilderValidators.compose([
-                    FormBuilderValidators.required(),
-                    FormBuilderValidators.minLength(2),
-                  ]),
-                  onSubmitted: _onSearchSubmitted,
-                ),
-              ),
+              const CompanySearchField(),
               Expanded(
                 child: BlocBuilder<SponsorCubit, SponsorState>(
                   builder: (context, state) {
